requirement.txt
langchain
langchain_core
python-dotenv
streamlit
langchain_community
